rename modules cgcnn prefix
