+++ cgcnn2/cli/cgcnn_pr.py
@@ -6,7 +6,9 @@ import warnings
 
 import numpy as np
 import torch
-from cgcnn2 import CIFData, CrystalGraphConvNet, cgcnn_test, collate_pool
+from cgcnn2.data import CIFData, collate_pool
+from cgcnn2.model import CrystalGraphConvNet
+from cgcnn2.utils import cgcnn_test
 from torch.utils.data import DataLoader
