+++ cgcnn2/cli/cgcnn_tr.py
@@ -6,8 +6,9 @@ from random import sample
 import numpy as np
 import torch
 import torch.nn as nn
-from cgcnn2 import (CIFData, CrystalGraphConvNet, Normalizer, cgcnn_test,
-                    collate_pool, get_lr, parse_arguments, train_force_split)
+from cgcnn2.data import CIFData, collate_pool
+from cgcnn2.model import CrystalGraphConvNet, Normalizer
+from cgcnn2.utils import cgcnn_test, get_lr, parse_arguments, train_force_split
 from sklearn.model_selection import train_test_split
 from torch.utils.data import DataLoader
