+++ cgcnn2/data.py
