+++ cgcnn2/utils.py
@@ -11,8 +11,8 @@ from pymatviz import density_hexbin
 from sklearn.metrics import mean_squared_error, r2_score
 from torch.utils.data import DataLoader
 
-from .cgcnn_data import CIFData_pred, collate_pool
-from .cgcnn_model import CrystalGraphConvNet, Normalizer
+from .data import CIFData_pred, collate_pool
+from .model import CrystalGraphConvNet, Normalizer
 
 
 def output_id_gen():
